Establish direct chat connections for redirects in Slack
Previously we were using Slack's `im.list` API method to look up
redirects to individual users. However, that method only returns
channels that the bot is currently participating in; if you try to
redirect to a user the bot hasn't interacted with before, you'd get a silent failure.

Now, we use the `im.open` API method, which idempotently establishes a
direct chat channel with a user, regardless of whether that user and the
bot have interacted before.

Additional refactoring as required to clean up and clarify.

Note: this only applies to the Slack adapter so far. Additionally, the
function signature of `lookup_room` can change now, but that has some
broader impacts, and should be addressed separately.