Instrument pipeline executor with event logging
This is the first step toward broader event logging across all of
Cog. For now, we add structured logging to the pipeline executor; each
place that logged before now uses a new `PipelineEvent` struct to
encapsulate details of events throughout the execution lifecycle. The
specifics of the event structure are detailed in the documentation for
`Cog.Events.PipelineEvent`. As we roll event logging to other parts of
the system, the `Cog.Events` namespace will get other event modules,
capturing details of other subsystems.

In brief, we log the start of the pipeline, each dispatch of a command
to a Relay, and the final success or failure of the pipeline.

In the future, we will log events to a file. For now, however, we use
the existing console logger; events will go to the same place as all
other log messages, but they will have a defined structure.

In the course of working through the executor code and making these
changes, a few other refactorings were made.

* The functionality of `send_to_command/1` has been merged into
`run_command/2` (and the former function removed entirely). This was
done in order to make it more clear what was happening, as well as
making relay information available for event logging without polluting
function signatures with either events or relay identifiers.

* `prepare_or_finish/2` was simplified as well. This function is only ever
called with a `resp.status` value of `"ok"`, so the case statement
pivoting on that value is unnecessary; the `"error"` clause was always
dead code.