Flag redirects to Slack channels the bot isn't in
Previously, if we attempted to redirect to a channel the bot wasn't a
member of, we would execute the pipeline but silently fail to
redirect. Slack does not allow posting to a channel the bot isn't in
via the RTM API.

(There are ways to use the HTTP API to do this, which we may do in the
future; we're not pursuing that now in order to prevent scope creep.)

Instead, we take advantage of the `is_member` key on Slack's channel
objects to determine whether the calling user (i.e., the bot) is a
member of the channel; if not, we flag this as a redirection error.

With this change, the channel caching behavior is changed slightly, in
that we only cache rooms the bot is a member of. This keeps the cache
small. We don't anticipate people trying to redirect to off-limits
rooms frequently (once you're told you're not a member, you'll either
stop trying to redirect there, or you'll invite the bot). In the
latter case, not caching the room when it's off-limits prevents us
from getting a stale cache hit that would block a subsequent retry
after inviting the bot; we'd hit the API again, see that we're in the
channel, and cache it then.

Other refactorings include extracting channel caching behavior to a
function to make it clearer exactly when and what is being cached, as
well as a handful of other small helper functions to clarify
intent. Documentation comments were also added, clarifying the details
of channel and user references that we receive from Slack, and how these
details affect our processing. We no longer cache the Slack room topic,
since we weren't doing anything with it any (it was leftover from
earlier days).

We're also not pulling back archived channels from the Slack API when
doing bulk channel listings; if we're not caching rooms the bot's not a
member of, we can also eliminate archived channels by definition. (In
the future, we may want to include them to indicate to the user that
they can't redirect because a channel is archived, but that can come
later.)

Error feedback to the chat user is given if a redirect is invalid due to
membership issues. It's not as clean as we'd like, but in the absence of
a general error templating solution, it at least provides actionable
feedback to the user.

Fixes #16