Save the execution context
We were pulling the value for cog_env off of the scope, but since we
collect lists of context when the command is of the 'once'
calling_convention, it makes cog_env difficult to use. This just
preserves the original output from the preceding command and passes it
along when required.