Move user resolution to head of executor flow
When emitting event logs for execution pipelines, it's better to include
the Cog user name as well as their adapter-specific chat handle (which
may be different!) in order to facilitate making system-wide
correlations. In order to do that, however, we need to move user
resolution to the very beginning of pipeline processing.

This ends up being nice, because if we determine that the invoking user
doesn't even have a Cog account, we can immediately abort all
processing; before, we did a fair amount of work before checking this.

Note: this change actually alters the behavior of the bot. Before, we
only resolved a Cog user if we needed to check permissions for the
command being run; if the command bypassed permission checks, we would
allow anybody to execute a command, even if they didn't have a Cog
account. This commit (by moving user resolution to the first processing
step, before a command has been parsed to see if it needs permission
checking) changes that, effectively requiring any user of the bot to
have a Cog account first.

The `PermissionInterpreter` no longer does the `UserPermissionsCache`
lookup, as this is now done by the executor directly. Additionally,
function heads, variable names, and pattern matches in
`UserPermissionsCache` were reworked to reflect the current state of the
cache (many reflected the time before it cached both users and
permissions).

Finishes work for #7