Bundle enabled/disabled: Toward saner delete!
Bundles now have an `enabled` field. In a nutshell, this determines
whether or not the bot will respond to commands from that bundle. This
can be used to quickly shut off command functionality, but (more broadly
speaking) is the first step toward a reliable bot-mediated command
uninstaller procedure; you'll shut off a bundle before you delete it.

REST API
========

Users may toggle a bundle between enabled and disabled with a REST API;
an example user flow might look like this:

```
GET /v1/bundles/ID/status
{"status":"enabled","relays":["44a92066-b1ae-4456-8e6a-4f212ded3180"],"bundle":"github"}

POST /v1/bundles/ID/status {"status": "disabled"}
{"status":"disabled","relays":["44a92066-b1ae-4456-8e6a-4f212ded3180"],"bundle":"github"}

GET /v1/bundles/ID/status
{"status":"disabled","relays":["44a92066-b1ae-4456-8e6a-4f212ded3180"],"bundle":"github"}

POST /v1/bundles/ID/status {"status": "enabled"}
{"status":"enabled","relays":["44a92066-b1ae-4456-8e6a-4f212ded3180"],"bundle":"github"}

GET /v1/bundles/ID/status
{"status":"enabled","relays":["44a92066-b1ae-4456-8e6a-4f212ded3180"],"bundle":"github"}
```

`POST` the desired state of the bundle to the endpoint and the API
handles the rest.

Changing the state of a bundle currently involves two operations: 1)
updating the database record for the bundle, and 2) notifying the
`Cog.Relay.Relays` server to update its bundle state appropriately. The
server uses this information to determine whether or not to return
relays for command dispatch.

Bundles may be enabled or disabled independently of whether any relays
happen to be serving them.

Refactorings
============

As part of the refactoring that took place here, a `Cog.Relay.Tracker`
module was extracted. This essentially abstracts the internal state of
the `Cog.Relay.Relays` server and encapsulates it with functions that
manipulate it.

Though it will only be used inside the `Relays` GenServer, splitting it
out has a few benefits:

* The internals of the GenServer are not cluttered with lots of data
  manipulation functions. You can focus solely on the "server-ness" of
  it now.
* It is easier to focus on the data manipulation. This makes it easier`
  to understand the data and the ins-and-outs of how it is
  processed. This should make any bugs easier to find.

Future Directions
=================

I went through several iterations on the way to the code in this
commit. At the end, I realized that a better way would be to have
`Cog.Relay.Relays` remain ignorant of bundle status, and instead make
the executor aware of it. The executor already needs to perform bundle
lookups as part of command disambiguation, so we could take advantage of
that and determine directly whether or not we should dispatch to a
relay.

However, this will require some surgery on the executor and command
disambiguation code, the former of which is already slated for some
refactorings soon. Given the nearness to launch, this change can wait
for later.