Extract bundle status manipulation into module
This consolidates logic for changing and displaying the status of a
bundle in one place, for use by the API and chat commands.

Model manipulation, `Cog.Relay.Relays` interaction, and embedded bundle
guards are all in one place now.

(Once `Relays` has no need to be aware of globabl bundle status, and the
choice of whether or not to dispatch commands is moved to the executor,
much of this can be simplified, moved back into `Cog.Models.Bundle`, or
removed entirely. Until then, it seemed prudent to isolate it.)