Depend explicitly on websocket_client
Recently we dropped an explicit dependency on `websocket_client`,
thinking instead that it would come from our Slack dependency. However,
it depends on the Github repository and not on a Hex package, so it
doesn't actually pull in `websocket_client` for us _at all_.

There is a `websocket_client` package in Hex, but it is from a fork of a
fork of the library we're currently depending on, and significant
changes have been made, causing it to not be a drop-in replacement.

Until we can sort this out properly, we'll depend on `websocket_client`
"manually" and carry on with life.