Overhaul `rules` command
Immediate goals are ensuring that all invocations return either
structured data that can be styled using templates, or error messages.

In the process, I imposed a more definite processing flow on the
`rules` command in an attempt to make it easier to maintain, and to
clearly show where various bits of processing occurred, and ensure
that error cases were being appropriately dealt with (e.g., invalid
option types, etc.). This brings the `rules` command more in line with
other commands like `permissions`.