Workaround for falling back to a default template
Previously, if a named template did not exist for the given bundle /
adapter combination, the `TemplateCache` would crash. Nothing would be
returned to the user, even though the pipeline had succeeded.

This commit introduces a workaround for this behavior; now, if the named
template cannot be found, we fall back to whatever the default rendering
of the response would be if no template had been specified.

Specifically, if the lookup fails, `TemplateCache.lookup/3` will return
`nil` instead of a rendering function. In this case, the `Executor` will
set the template to `nil` and run it through the rendering logic again
to pick up the default.

Note that this is NOT a long-term solution; the true fix will involve a
consolidation of templating logic that is currently spread across the
`TemplateCache` and the `Executor`, as well as a simplification /
clarification of the template lookup "chain".

Fixes #232