Bundles are disabled by default
When installing a bundle, it should always be disabled until explicitly
enabled by an administrator.

Here, we add a migration that changes the database default, as well as
the bundle model default.

Additionally, the embedded bundle is the only one that should be
installed as enabled (rather hard to do anything without _that_). A
changeset function has been added to the bundel model to ensure that the
embedded bundle is always enabled.

Fixes #124