Create adapter process to talk to emqtt
A new `Adapter` module was added to handle all emqtt communication. Each
service creates its own adapter that uses this module and implements
the adapter behaviour. The behaviour also changed a bit to reply on a
supervisor module instead of describing the tree of processes, along
with some minor function name changes.