Database is sole source of truth for bundle status
Previously, `Cog.Relay.Relays` was tracking not only which relays were
running a given bundle, but also the enabled/disabled status of those
bundles. Now, it tracks only which relays run which bundles, and knows
nothing of the bundles' status.

Instead, the executor now factors in the bundle status when
determining which relay to dispatch to. If the bundle is disabled, we
can now return a more accurate error message to the user (previously,
it was confusing to say that no relays were running a bundle if some
relays actually _were_, but the bundle was just disabled).

To ease the database burden for these status checks, a new bundle
cache is introduced. It is similar to other caches we have, but
implemented to use generic functions as much as possible. Hopefully
this can be the basis of a future generic cache behaviour that we can
use elsewhere.