New and improved executor
Uses the new parser and simplifies A LOT of what the executor does.

Introduces a new concept of an "execution plan" for pipeline stages,
in an attempt to decouple from the parser AST (though that's not quite
complete just yet).

Lots more remains to be done, but this takes care of a huge chunk of
the work.