Decouple room lookup and membership check
This commit separates room membership checks from room lookups. This is
achieved by adding a new public API function to the `Cog.Adapter`
behavior: `room_writeable?/1` which returns true when the bot is
permitted to send messages to a room, false when it's not, and an error
tuple for all other cases. All adapters except Slack always return true.

The Slack adapter returns true if the bot is a member of the room
otherwise false (or error if the API call fails). This allows us to
catch the case where a user redirects output to room(s) where their Cog
instance is not present.

The executor uses this check to detect invalid redirects.

Fixes #410