Pipeline ID can be passed into the executor
Previously the executor itself would create a unique ID for the pipeline
being executed. In order to facilitate better tracking for pipelines
that are initiated from external events (e.g., webhooks), we'll need to
externalize this ID generation. (this will allow us to correlate HTTP
request processing with the resulting pipeline).

Now the pipeline initializer creates the ID and passes it as an argument
to the new executor process.

Fixes #417