Allow an initial context to be passed to Executor
Previously, we had a hard-coded initial context inside the Executor. For
pipelines triggered by external events, we'll need a mechanism by which
to seed that initial context with arbitrary data (e.g., webhook request
bodies).

This lays the groundwork for that: if a request comes in with a map or
list stored under the "initial_context" key, that will be used. If
nothing is explicitly set, the same initial context as before (namely,
an empty map) will be used. Thus, all chat-adapter initiated pipelines
will continue to behave as before.

Created `store_with_template/2` and `remove_template/1` helper functions
to better encapsulate some manipulation operations so they're more
obvious.

Fixes #418