Target redirects to specific adapters
This is groundwork for upcoming work to trigger pipelines in response to
external events (e.g., webhooks). In such a case, a pipeline will be
triggered from one adapter (e.g., an "HTTP adapter"), but output will
need to be directed to another (namely, the chat adapter).

Here, we introduce an optional URL-style notation for specifying
redirect targets. By redirecting to `chat://#room`, a response will be
formatted and delivered via the currently-configured chat adapter. The
redirect target itself will be resolved according to the chat adapter,
as well. Redirects may continue to be specified without the URL
prefix (e.g., `#room`), in which case they will be resolved and
dispatched according to the adapter the request originated from.

The special `me` and `here` redirect targets are always resolved
relative to the originating adapter.

As part of this work, the reply topic of an adapter is now exposed as a
behaviour function, and the executor discovers the appropriate reply
topics to send responses to, rather than relying on whatever is sent in
the original request.

This change necessitated a modification to how we test
adapters. Previously, we created a unique reply topic for each message
and waited to receive a message on that topic. Since we no longer look
at any reply topic sent to the executor, but rather ask adapter modules
directly what topics they listen on, this broke several tests. Each
pipeline executor process is given a unique ID from the adapter now,
however, which we now incorporate into the outgoing response message. By
matching this value, we can salvage the adapter test helper.

Additionally, the `ExecutorHelpers` module was removed, its functions
absorbed into the executor in order to consolidate and unify how
messages are sent out from the executor. Further streamlining of error
handling is currently scheduled.

Fixes #415