Overhaul executor error-handling
Previously, error messages were generated in a number of places, making
it difficult to get a grasp on the variety of responses generated.

Additionally, error messages did not contain all that much context about
the error, something we'd need to address before allowing pipelines to
be triggered by external events.

Here, (almost) all error handling is consolidated into one place in the executor,
with lots of additional context added.

(Error messages for people lacking a Cog account, as well as for bad
redirects, are still handled as they were, since these were already
complex messages. Once we have error templates, everything will be able
to be consolidated even further.)

Tests have been updated, but in an admittedly less-than-satisfying
way. They are still strongly coupled to the exact output of the error
messages, but upcoming work to template error responses will,
interestingly enough, allow us to perform comparisons on data
structures, and not text, which will make the tests more robust in the
face of future changes.

Fixes #416