Trigger pipelines in response to external events
Now pipelines may be triggered by HTTP requests instead of just via chat
messages. This will allow Cog to, e.g., respond to webhooks.

Users may define "event hooks", which define a pipeline to run in
response to an HTTP POST request. The request itself acts as input to
the pipeline (a COG_ENV for the first command in a pipeline,
technically), allowing query parameters, request headers and JSON
request bodies to be accessible to the command. Pipelines may run as a
defined user, or may require token-based authentication for
execution. Additionally, pipelines may send complete output back to the
requestor via HTTP response, or direct output to the currently-enabled
Cog chat adapter, or both.

A "CRUD" REST API is provided for event hooks and is currently the sole
means of manipulating event hooks (i.e., no chat commands for this just
yet). Execution of hooks, while done via HTTP request, is handled by a
different Phoenix Endpoint than the REST API (defaulting to
port 4001, vs. the REST API's default port 4000).

To do this, we introduce a new "HTTP adapter", which is an
implementation of our current Adapter behaviour. This adapter is
different than previous ones in a few ways, though.

First, it is a synchronous adapter. All chat adapters thus far have been
asynchronous, in that responding to a Cog request does not need to be
coupled to the _receiving_ of that request. As we are now responding to
HTTP requests, this is no longer the case; we _must_ send a response
when we receive a request! To deal with this, an "AdapterBridge" is
introduced to track which pipeline executions correspond to which
currently-processing HTTP request. (If you squint, this "AdapterBridge"
performs the same function for the HTTP adapter as our Slack RTM/API and
HipChat API modules perform for those adapters.)

Secondly, as a synchronous adapter, the HTTP adapter is the first to
send output to mulitple places _via multiple adapters_. As mentioned, it
must _always_ send something back for the HTTP response. However, it
must _also_ be able to (optionally) send output to chat.

Third, the HTTP adapter is the first to seed a pipeline with data; a
`COG_ENV` is now presented to the pipeline from the beginning, rather
than being generated during execution. This allows commands access to
the data in the HTTP request. To facilitate correlation of HTTP request
processing and resultant pipeline execution, the HTTP adapter also
passes the unique ID of the HTTP request to the Executor as the
pipeline's unique ID. As a result, the Adapter behaviour contract was
modified to allow passing these two pieces of data in with a request.

To help manage these differences in a general way, adapters now know
whether they are a chat adapter or not, with the HTTP adapter being the
first non-chat adapter. Behavior in the Executor and associated code can
ask the adapter which it is in order to do the right thing.

Other changes of note:

* The `NOCHAT` behavior has been dropped. If you want to run without a
  chat adapter, choose the `null` adapter.
* A new embedded permission `operable:manage:hooks` has been added; all
  CRUD REST API interaction around event hooks requires this permission.
* The `TemplateCache` has been moved from the top-level supervisor down
  to the `CommandSup` supervisor.
* Pipeline destination processing grew more complicated, now that output
  can be "full" or "status-only" (the latter for HTTP request that don't
  redirect output to `here`, but still require something for the HTTP
  response), so a new `Cog.Command.Pipeline.Destination` module was
  created to manage destination processing. This removes a fair bit of
  code from the Executor, as well as an entire FSM state.
* Lots of tests
* A new testing utility `Cog.Snoop` is used to eavesdrop on MQTT message
  traffic. Here it is used to verify certain aspects of event hook
  processing, but may be useful in other tests,as well.

Closes #344
Closes #385
Closes #387