Timeouts now return 202 instead of 500
If a pipeline takes longer than an event hook's specified timeout, it
will return a 202 now instead of a 500, to reflect the fact that the
request was successfully received and processed, but that it just hasn't
finished fully executing.

To help test this, a new `sleep` command was added to the embedded
bundle. It pauses execution of the pipeline for a configurable number of
seconds before passing the entire COG_ENV through unchanged.