Manually parse JSON body in hook requests
We'd like to hang on to the raw body text to allow users to perform
validation based on the initial request (e.g., verifying header-provided
signatures based on a shared secret).

We have to provide the raw text because we can't recapitulate it from
our parsed JSON because maps don't have a defined order. In order to get
the raw text, though, we have to bypass Phoenix's built-in parsers.