Provide binary-only interfaces for lookup_user
I believe that the keyword argument version of lookup_user is only used
internally to the various adapters, and that the `id: id` argument is no
longer used.

In the near future, we should split apart the id and handle versions of
this function, and remove or limit the scope of the id version as
appropriate.