Relay installs embedded bundle only
Changed lib/cog/relay/relays.ex to only install the embedded bundle. If
it receives an announcement for an unknown bundle other than the
embedded bundle it will log an error and return failed to the announcing
relay.

Also moved where we log the removal of a Relay to avoid a misleading
"removed relay xyz" message from getting logged when, in fact, all Cog
has done is updated the list of bundles known by a particular Relay.