Incorporate a buffer to trigger timeouts
Trigger timeouts are defined according to the needs of the
requestor (e.g. Github webhooks must respond within 30 seconds), which
includes network roundtrip time, as well as Cog's internal
processing. Cog itself can't wait that long to respond, as that'll be
guaranteed to exceed the HTTP requestor's timeout. As such, we'll
incorporate a buffer into our internal timeout, allowing us to return to
the requestor within their timeout.

This can be configured on a system-wide basis.