Create example `docker-compose.yml` file
Instead of trying to provide a "one size fits all"
`docker-compose.yml` file, we instead provide an minimal, explicit
example file with instructions on how to customize it.

We also hard-code a tag for the cog image to use.

Real `docker-compose.yml` files, as well as `.env` files, are now
ignored by git.