Chat handle creation is now an "upsert"
Previously it was a bit awkward to know whether or not you needed to
create or update a chat handle for a user. It turns out we can simplify
things, because a user can only ever have a single chat handle for a
given chat provider. Whether the user already has a chat handle or not,
we can just say we're "setting" the chat handle and basically do an
"upsert" operation on the backend.

This removes the `PUT` operation on the chat handle controller. Now, to
both create and update a user's chat handle, send a `POST` to
`/v1/users/:id/chat_handles`.

This also puts a guard in against the situation where a user tries to
set a chat handle for a provider that differs from the
currently-configured Cog chat adapter. Previously, this would crash Cog;
now it's just an error tuple.

This also pulls this chat handle logic out of the controller and into a
chat handle "repository", in keeping with the new work to consolidate
domain functionality in the core of the system.

Fixes #604