Nest chat_handles under users
We needed a way for non-admin users to view their own chat handles. We
could have opened up chat_handle#index to the current_user, but I think
this is a more desirable option since it's a bit easier from an
implementation side in Cog, and it prevents having to make a second
request just to get the associated chat handle.