Manage rule deletion and creation
Keeping track of which rules are present in various bundle versions is
important for our upgrade process, as well as for how the end user will
manage their rules. If a rule that was introduced in version X of some
bundle is not to the user's liking, they should be able to remove it,
and have it stay gone even if version X+1 of the bundle re-introduces
that same rule.

To achieve this, "rule deletion" is usually "rule disabling", behind the
scenes. If a rule is associated with a non-"site" bundle version, the
rule is not truly deleted, but is rendered inactive. If the rule is only
present in the "site" bundle (that is, the end user entered it
themselves), a deletion will actually remove it from the system. In the
event that a user has disabled a bundle rule but then subsequently
enters in a new "site" rule with the same rule text, we'll link the rule
to the the site bundle and re-enable it. Deleting this rule will return
it to its original disabled state.

Enabling / disabling is not intended to be user-visible, but is an
implementation detail.