Override service and trigger base URLs
This commit introduces two new environment variables which can be used
to override the base URLs for services and triggers:

$COG_TRIGGER_URL_BASE
$COG_SERVICE_URL_BASE

Setting either/both of these variables will override the base portion of
the corresponding. This capability is useful when Cog is unable to
detect the correct routable URL to send to clients. For example,
firewalls and load balancers interpose themselves between clients and
servers and require hosts behind them to carefully build any URLs
returned in output so clients are able to access those resources.

Fixes #694