Add uniqueness constraint to bundle version model
This prevents crashes when attempting to install the same bundle version
twice (as determined by bundle name and version string _only_; it
doesn't care if the contents of the bundle configuration file are
identical between versions).

We wrap all the bundle installation logic in a try/rescue block, and
trigger a transaction rollback if any `Ecto.InvalidChangesetError`s are
raised. Thus, this will actually protect against _any_ error raised in
the installation process.