First stab at handling overly large chat responses
Chat adapters can have limits on the size of message they will
accept. While a complete solution will deal with this in an
adapter-specific manner (breaking long messages into multiple messages,
using an alternative format, etc.), we can at least log a message when
we see a message that is potentially "too big".

Here, we add a new option recognized by
`Carrier.Messaging.Connection.publish/3` that lets a caller specify a
size threshold in bytes. If the JSON-encoded message is larger than this
threshold, we emit a log warning.

Fixes #739