Retry failed room lookups w/common variants
This is a incredibly hacky fix to address Cog's current difficulties
interacting with capitalized HipChat rooms. The core issue is the HipChat
send message API requires a room ID. Normally we obtain this ID via HipChat's
API to look up room IDs from room names.

The complicating bit is that the room name sent to us in the XMPP message is
downcased which causes the ID look up to fail. There appears to be no other
useful information embedded in the XMPP message which gives us enough information
to successfully look up the room.

I suspect the correct long term solution here is to see if we can move the bulk
of our interactions w/HipChat to purely XMPP. We could then avoid these types
of impedance mismatches. In the short term, this commit implements a complete hack
which hopefully enables the majority of users hitting this issue to work. To wit,
when the HipChat adapter fails to look up a room by name it retries up to two more
times before giving up. The first retry capitalizes the room name. If that lookup
fails the adapter uses the `Inflex` library to camel case the room name. If neither
look ups succeed then an error is returned.