Query rule for mentioned permissions
This commit uses the newly added Permissions.Ast.Rule.permissions_used/1
function to obtain the list of permissions mentioned in a access
rule to generate a better access denied error message. This also
prevents Cog from crashing when rules mention more than one permission.

This isn't an ideal solution as we're not telling the user exactly which
specific permissions they're lacking but it is a step in the right
direction and vastly better than crashing.

50% fix for #758