Randomly select Relay at beginning of stage only
This commit modifies the Relay selection logic in Cog.Executor such that
a Relay is randomly selected once at the start of each pipeline stage
and is then reused for subsequent invocations for that stage. Doing so
allows Cog to take maximum advantage of Relay's ability to reuse
containers for the same pipeline stage and is expected to improve
performance for Cog installations with multiple Relays.