Move Relay selection to the planner.
Previous iteration assigned Relays when each plan was executed. This was
wrong as a single invocation can generate multiple plans which negated
the benefits of selecting a Relay when each plan executes.

This commit moves Relay selection into the planner. A Relay is selected
before plans for an invocation are created and shared across all the
created plans.