Cache and reuse relay assignments
This commit modifies Cog.Command.Pipeline.Executor to remember and reuse
the Relay chosen to run an invocation for a bundle version. The memory
persists between plans and stages allowing Cog to take maximum advantage
of Relay's container reuse for a given pipeline.

The Executor also performs a safety check when it uses a cache Relay
assignment. Before executing `current_plan`, the Executor double checks
the selected Relay is still valid for the bundle version. If the Relay
assignment is invalid (Relay is offline or the bundle version has been
disabled/uninstalled) the Executor will attempt to select another Relay
once before giving up and returning an error to the user.