Add a `relay-group list` subcommand
Turns out the `relay-group info` command behaved like this when given 0
arguments; this basically pulls that out into a `list` command,
consistent with other commands.

This new `list` subcommand is not the default if a subcommand is not specified.

(While I was here, I fixed the `relay-group-info-verbose` template,
which emitted a `Bundles:` line for each bundle, prior to listing the
names of all bundles at the end)