Squat on the "cog" bundle name
While we don't currently have a "cog" bundle, we may provide one in the
future. To preserve for that possibility, and to prevent possible
confusion with 3rd party bundles that might arise, we're going to
prevent bundles named "cog" from being installed for now.

We were using "cog" as a bundle name for testing data; obviously that's
out now, so we'll now call it "test-bundle" instead.

Fixes #817