Enable user self-registration
This commit adds support for creating new users upon their first
interaction with Cog. If the user is unknown to Cog it will create a
corresponding user and chat handle records before allowing the request
to proceed.

This behavior is controlled by a new `:cog` config entry named
`:self_registration` which defaults to false (disabled).