Automatically manage embedded bundle version
We're going to start to tease apart the embedded bundle from the core
Cog application by separating their version numbers from each
other. Right now, it's not possible to fully break them apart due to how
intertwined the bundle and the core application code are.

The main reason for doing this is to give us a bit of leeway for rolling
back Cog installations if a nasty bug shows up. Previously, the embedded
bundle would block any attempt at a rollback, because it can't be rolled
back itself. Now, with the versions decoupled, we can have a little bit
of wiggle room. In an ideal world, we would always roll forward when
such a bug was found, but this gives us some options we wouldn't
otherwise have.

This commit introduces a Mix task that runs automatically after
compilation, checksumming the compiled code and template files of the
embedded bundle and automatically bumping a configuration variable
containing the bundle's version string if it detects any changes. All
the developer needs to do is commit the updated information. Note that
the developer still has discretion over whether to "really" bump the
version number. For instance, if there are indeed code changes, but
they're refactoring changes that don't alter behavior or the underlying
data model, we can just commit the updated checksum, but not the updated
version.

Overall, this is still a bit of a hack until we can fully extricate the
embedded bundle (e.g., by making `cogctl` able to run as a bundle
itself).

VERY IMPORTANT: In order for this Mix task to work properly, you MUST be
running Elixir 1.3.1 or later (might work on 1.3.0, but hasn't been
tried). Earlier versions of the Elixir compiler appear to generate BEAM
files in a non-deterministic manner (particularly when macros are
involved), yielding different checksums across compilations of the exact
same source code. If you start seeing the bundle version incrementing
each time you compile, you're probably running Elixir 1.2.x and should
upgrade ASAP.