Add custom EEx template rendering engine
This adds  an alternative  EEx engine  that processes  templates while
restricting the  amount of Elixir  and Erlang  that can be  called. It
also provides a way to expose custom helper functions in the execution
context.

This will allow us to transition from Mustache templates. We'll now be
able to provide helpers, and also be able to add a bit of basic logic
to templates (we can have `if` statements!)

This is not currently hooked up to the rest of Cog yet; templates
continue to be Mustache for now.