Only checksum embedded bundle when MIX_ENV=dev
TL;DR - checksums can change with different MIX_ENV, a detail that is
irrelevant for our purposes. Read the code comments for details.

Also changes the stored checksum for the embedded bundle; the old value
was actually computed under MIX_ENV=test. This change returns it to the
MIX_ENV=dev checksum that it originally was (but the version will stay
the same, because that's how this stuff works....)