Support layered dynamic bundle configuration
Adds the ability to manage "layers" of dynamic bundle configuration. The
previous notion of dynamic configuration is referred to now as the
"base" layer, but now we may add "room" and "user" layers on top. Relay
can apply these in order to override the base layer depending on which
user issues a command, and from which room it is issued.

Layers can have names, indicating which user or room they apply to. The
base layer is always named "config" by convention (it could be named
anything, really).

There can only be one base layer per bundle, but multiple room and user
layers, provided they each have a different name.

A REST API is provided to manipulate each layer of a bundle
individually. Additionally, all layers for a bundle may be retrieved at
once.

Cog now returns all layers for all bundles that a Relay is currently
serving, grouped by bundle. This allows us to optimize processing on
Relay. (We also reduce the amount of information we send, as we don't
need each layer to have the bundle name, or the checksum hash.)